fixed logic

+++ jmp-cloud-service-impl/src/main/java/com/epam/learn/service/subscription/DefaultSubscriptionDtoService.java
@@ -11,31 +11,46 @@ import com.epam.learn.dto.SubscriptionRequestDto;
 import com.epam.learn.dto.SubscriptionResponseDto;
 import com.epam.learn.entity.Subscription;
 import com.epam.learn.exception.NotFoundException;
-import com.epam.learn.service.converter.SubscriptionRequestDtoToSubscriptionConverter;
-import com.epam.learn.service.converter.SubscriptionToSubscriptionResponseDtoConverter;
-import com.epam.learn.subscription.SubscriptionDtoConverterService;
+import com.epam.learn.subscription.SubscriptionDtoService;
 import com.epam.learn.subscription.SubscriptionService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import static java.lang.String.format;
 
+/**
+ * Service that works with {@link SubscriptionRequestDto} and {@link SubscriptionResponseDto}
+ */
 @Service
-public class SubscriptionDtoConverterServiceImpl implements SubscriptionDtoConverterService {
+public class DefaultSubscriptionDtoService implements SubscriptionDtoService {
 
     private final SubscriptionService subscriptionService;
     private final SubscriptionToSubscriptionResponseDtoConverter subscriptionToSubscriptionResponseDtoConverter;
     private final SubscriptionRequestDtoToSubscriptionConverter subscriptionRequestDtoToSubscriptionConverter;
 
-    public SubscriptionDtoConverterServiceImpl(SubscriptionService subscriptionService, SubscriptionToSubscriptionResponseDtoConverter subscriptionToSubscriptionResponseDtoConverter, SubscriptionRequestDtoToSubscriptionConverter subscriptionRequestDtoToSubscriptionConverter) {
+    public DefaultSubscriptionDtoService(SubscriptionService subscriptionService, SubscriptionToSubscriptionResponseDtoConverter subscriptionToSubscriptionResponseDtoConverter, SubscriptionRequestDtoToSubscriptionConverter subscriptionRequestDtoToSubscriptionConverter) {
         this.subscriptionService = subscriptionService;
         this.subscriptionToSubscriptionResponseDtoConverter = subscriptionToSubscriptionResponseDtoConverter;
         this.subscriptionRequestDtoToSubscriptionConverter = subscriptionRequestDtoToSubscriptionConverter;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     @Transactional
-    public SubscriptionResponseDto saveOrUpdateSubscription(SubscriptionRequestDto subscriptionRequestDto) {
+    public SubscriptionResponseDto saveSubscription(SubscriptionRequestDto subscriptionRequestDto) {
+        Subscription subscription = subscriptionRequestDtoToSubscriptionConverter.convert(subscriptionRequestDto);
+        subscription = subscriptionService.saveOrUpdateSubscription(subscription);
+        return subscriptionToSubscriptionResponseDtoConverter.convert(subscription);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    @Transactional
+    public SubscriptionResponseDto updateSubscription(SubscriptionRequestDto subscriptionRequestDto) {
         Long subscriptionRequestDtoId = subscriptionRequestDto.getId();
         Subscription subscription = Optional.ofNullable(subscriptionRequestDtoId).map(id -> updateExistingSubscription(subscriptionRequestDto, id))
                 .orElse(subscriptionRequestDtoToSubscriptionConverter.convert(subscriptionRequestDto));
@@ -43,14 +58,20 @@ public class SubscriptionDtoConverterServiceImpl implements SubscriptionDtoConve
         return subscriptionToSubscriptionResponseDtoConverter.convert(subscription);
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     @Transactional(readOnly = true)
     public SubscriptionResponseDto getSubscription(Long id) {
-        Subscription subscription = getSubscriptionFromRepo(id)
+        Subscription subscription = fetchSubscription(id)
                 .orElseThrow(() -> new NotFoundException(format("Subscription with provided id: %s cannot be found", id)));
         return subscriptionToSubscriptionResponseDtoConverter.convert(subscription);
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     @Transactional(readOnly = true)
     public List<SubscriptionResponseDto> getAllSubscription() {
@@ -59,14 +80,29 @@ public class SubscriptionDtoConverterServiceImpl implements SubscriptionDtoConve
                 .collect(Collectors.toList());
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     @Transactional
     public void deleteSubscription(Long id) {
-        subscriptionService.deleteSubscription(id);
+        if (subscriptionExists(id)) {
+            subscriptionService.deleteSubscription(id);
+        } else {
+            throw new NotFoundException(format("Subscription with provided id: %s cannot be found", id));
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean subscriptionExists(Long id) {
+        return subscriptionService.subscriptionExists(id);
     }
 
     private Subscription updateExistingSubscription(SubscriptionRequestDto subscriptionRequestDto, Long id) {
-        return getSubscriptionFromRepo(id)
+        return fetchSubscription(id)
                 .map((subscriptionEntity) -> {
                     subscriptionRequestDtoToSubscriptionConverter.updateExistingSubscription(subscriptionRequestDto, subscriptionEntity);
                     return subscriptionEntity;
@@ -74,7 +110,7 @@ public class SubscriptionDtoConverterServiceImpl implements SubscriptionDtoConve
                 .orElseThrow(() -> new NotFoundException(format("Subscription with provided id: %s cannot be found", id)));
     }
 
-    private Optional<Subscription> getSubscriptionFromRepo(Long id) {
+    private Optional<Subscription> fetchSubscription(Long id) {
         return subscriptionService.getSubscription(id);
 
     }
