+++ jmp-cloud-service-impl/src/main/java/com/epam/learn/service/user/DefaultUserDtoService.java
@@ -11,38 +11,55 @@ import com.epam.learn.dto.UserRequestDto;
 import com.epam.learn.dto.UserResponseDto;
 import com.epam.learn.entity.User;
 import com.epam.learn.exception.NotFoundException;
-import com.epam.learn.service.converter.UserRequestDtoToUserConverter;
-import com.epam.learn.service.converter.UserToUserResponseDtoConverter;
-import com.epam.learn.user.UserDtoConverterService;
+import com.epam.learn.user.UserDtoService;
 import com.epam.learn.user.UserService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import static java.lang.String.format;
 
+/**
+ * Service that works with {@link UserRequestDto} and {@link UserResponseDto}
+ */
 @Service
-public class UserDtoConverterServiceImpl implements UserDtoConverterService {
+public class DefaultUserDtoService implements UserDtoService {
 
     private final UserService userService;
     private final UserToUserResponseDtoConverter userToUserResponseDtoConverter;
     private final UserRequestDtoToUserConverter userRequestDtoToUserConverter;
 
-    public UserDtoConverterServiceImpl(UserService userService, UserToUserResponseDtoConverter userToUserResponseDtoConverter, UserRequestDtoToUserConverter userRequestDtoToUserConverter) {
+    public DefaultUserDtoService(UserService userService, UserToUserResponseDtoConverter userToUserResponseDtoConverter, UserRequestDtoToUserConverter userRequestDtoToUserConverter) {
         this.userService = userService;
         this.userToUserResponseDtoConverter = userToUserResponseDtoConverter;
         this.userRequestDtoToUserConverter = userRequestDtoToUserConverter;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     @Transactional
-    public UserResponseDto saveOrUpdateUser(UserRequestDto userRequestDto) {
-        Long userRequestDtoId = userRequestDto.getId();
-        User user = Optional.ofNullable(userRequestDtoId).map(id -> updateExistingUser(userRequestDto, id))
-                .orElse(userRequestDtoToUserConverter.convert(userRequestDto));
+    public UserResponseDto saveUser(UserRequestDto userRequestDto) {
+        User user = userRequestDtoToUserConverter.convert(userRequestDto);
         user = userService.saveOrUpdateUser(user);
         return userToUserResponseDtoConverter.convert(user);
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    @Transactional
+    public UserResponseDto updateUser(UserRequestDto userRequestDto) {
+        Long userRequestDtoId = userRequestDto.getId();
+        User user = Optional.ofNullable(userRequestDtoId).map(id -> updateExistingUser(userRequestDto, id))
+                .orElseThrow(() -> new NotFoundException(format("User with provided id: %s cannot be found", userRequestDtoId)));
+        return userToUserResponseDtoConverter.convert(user);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
     @Override
     @Transactional(readOnly = true)
     public UserResponseDto getUser(Long id) {
@@ -51,6 +68,9 @@ public class UserDtoConverterServiceImpl implements UserDtoConverterService {
         return userToUserResponseDtoConverter.convert(user);
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     @Transactional(readOnly = true)
     public List<UserResponseDto> getAllUser() {
@@ -59,10 +79,26 @@ public class UserDtoConverterServiceImpl implements UserDtoConverterService {
                 .collect(Collectors.toList());
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     @Transactional
     public void deleteUser(Long id) {
-        userService.deleteUser(id);
+        if (userExists(id)) {
+            userService.deleteUser(id);
+        } else {
+            throw new NotFoundException(format("User with provided id: %s cannot be found", id));
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    @Transactional
+    public boolean userExists(Long id) {
+        return userService.userExists(id);
     }
 
     private User updateExistingUser(UserRequestDto userRequestDto, Long id) {
